@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -10,21 +10,14 @@ import { GlobalModule } from './modules/global/global.module';
 import { HomePageModule } from './modules/home-page/home-page.module';
 import { CarouselModule } from 'ngx-bootstrap/carousel';
 
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    BrowserAnimationsModule,
-    GlobalModule,
-    FormsModule,
-    HomePageModule,
-    CarouselModule,
-    HttpClientModule
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        GlobalModule,
+        FormsModule,
+        HomePageModule,
+        CarouselModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
